% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exdat.R
\docType{data}
\name{exdat}
\alias{exdat}
\title{Example data based on Takeya (1991)}
\format{A data frame with 10 rows and 5 variables}
\usage{
exdat
}
\description{
A dataset containing 10 observations on 5 items.
}
\keyword{datasets}

