% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scatterplot.R
\name{scatterplot}
\alias{scatterplot}
\title{Scatterplot Matrices}
\usage{
scatterplot(data, select = NULL, type = c("jitter", "size", "count", "sun"))
}
\arguments{
\item{data}{a data frame}

\item{select}{select items to be plotted}

\item{type}{type of plot, i.e., 'jitter', 'size', 'count', and 'sun'}
}
\description{
This function produces a scatterplot matrix
}
\details{
Using a scatterplot matrix, an overview of the answer patterns for the pairs of items can be taken.
}
\examples{
# Example data based on Takeya (1991)

# Select items to be plotted
scatterplot(exdat, select = c("Item2", "Item3", "Item4"))

# Scatterplot matrix: jitter
scatterplot(exdat)

# Scatterplot matrix: size
scatterplot(exdat, type = "size")

# Scatterplot matrix: count
scatterplot(exdat, type = "count")

# Scatterplot matrix: sun
scatterplot(exdat, type = "sun")
}
\author{
Takuya Yanagida \email{takuya.yanagida@univie.ac.at},
Keiko Sakai \email{keiko.sakai@oit.ac.jp}
}
\references{
Takeya, M. (1991). \emph{A new test theory: Structural analyses for educational information}.
Tokyo: Waseda University Press.
}
\seealso{
\code{\link{TSSA}}, \code{\link{SSRA}}
}

