% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scatterplot.R
\name{scatterplot}
\alias{scatterplot}
\title{Scatterplot Matrices}
\usage{
scatterplot(data, type = c("jitter", "size", "count", "sun"))
}
\arguments{
\item{data}{a data frame}

\item{type}{type of plot, i.e., 'jitter', 'size', 'count', and 'sun'}
}
\description{
This function produces a scatterplot matrix
}
\details{
Takea Semantic Structure Analysis (TSSA) and Sakai Sequential Relation Analysis (SSRA) are graphical approaches
}
\examples{
# Example data based on Takeya (1991)

# Scatterplot matrix: jitter
scatterplot(exdat)

# Scatterplot matrix: size
scatterplot(exdat, type = "size")

# Scatterplot matrix: count
scatterplot(exdat, type = "count")

# Scatterplot matrix: sun
scatterplot(exdat, type = "sun")
}
\author{
Takuya Yanagida
Keiko Sakai
}
\references{
Takeya, M. (1991). \emph{A new test theory: Structural analyses for educational information}.
Tokyo: Waseda University Press.
}
\seealso{
\code{\link{TSSA}}, \code{\link{SSRA}}
}

